refactor(centroid): deduplicate shared stages and split into a module dir
Tier 5 centroid-extraction compaction (C8, C9, C11/C12 parts):

Dedup (C8) — five shared helpers now serve both extraction paths:
- midpoint_f32 / median_f32 replace five hand-rolled median idioms (the
  block-medians site also gains O(n) selection over its former full sort)
- sort_and_truncate_by_mass replaces the two identical brightest-first tails
- check_pixel_len replaces the two identical length validations
- accepted_peak_refine replaces the duplicated 3x3-parabola gate (npix >= 5,
  border check, 0.5 px CoM-agreement) around quadratic_peak_offset

Split (C9) — centroid_extraction.rs (1800 lines) becomes
centroid_extraction/{mod,ccl,fast}.rs: crate-facing types, entry points and
shared helpers in mod.rs; the CCL pipeline (incl. its parallel dispatch
module) in ccl.rs; the single-pass tracker path in fast.rs. No API changes;
FastCentroidConfig / extract_centroids_fast re-exported as before.

Also:
- extract_from_gray borrows the input via Cow when local_bg_block_size is
  None instead of making two full-image copies (C11)
- num_blobs_raw on the CCL path now reports connected components before the
  size/elongation filters, matching its documented meaning and the fast
  path's semantics (was: the post-filter count)
- fixed the stale "normalize to [0, 65535]" comment (values are cast)

All feature combos build (image, image+parallel); lib + skyview + TESS +
integration + Python suites pass.
